+++ python/jaeger/commands/status.py
@@ -6,7 +6,6 @@
 # @Filename: status.py
 # @License: BSD 3-clause (http://www.opensource.org/licenses/BSD-3-Clause)
 
-
 import numpy
 
 from jaeger.commands import Command, CommandID
@@ -23,6 +22,7 @@ class GetID(Command):
     command_id = CommandID.GET_ID
     broadcastable = True
     timeout = 1
+    safe = True
 
     def get_ids(self):
         """Returns a list of positioners that replied back."""
@@ -35,6 +35,7 @@ class GetStatus(Command):
 
     command_id = CommandID.GET_STATUS
     broadcastable = True
+    safe = True
 
     def get_positioner_status(self):
         """Returns the `~.maskbit.PositionerStatus` flag for each reply."""
@@ -48,6 +49,7 @@ class GetActualPosition(Command):
 
     command_id = CommandID.GET_ACTUAL_POSITION
     broadcastable = False
+    safe = True
 
     def get_positions(self):
         """Returns the positions of alpha and beta in degrees.