+++ python/jaeger/core/exceptions.py
@@ -17,10 +17,13 @@ class JaegerError(Exception):
 
     def __init__(self, message=None):
 
-        message = 'There has been an error' \
-            if not message else message
+        message = 'There has been an error' if not message else message
 
         super(JaegerError, self).__init__(message)
+
+
+class FPSLockedError(JaegerError):
+    """The FPS is locked."""
 
 
 class JaegerNotImplemented(JaegerError):