+++ python/jaeger/maskbits.py
@@ -82,10 +82,19 @@ class PositionerStatus(Maskbit):
     DISPLACEMENT_COMPLETED = 0x01000000
     ALPHA_DISPLACEMENT_COMPLETED = 0x02000000
     BETA_DISPLACEMENT_COMPLETED = 0x04000000
+    ALPHA_COLLISION = 0x08000000
+    BETA_COLLISION = 0x10000000
     DATUM_INITIALIZED = 0x20000000
     ESTIMATED_POSITION = 0x40000000
     POSITION_RESTORED = 0x80000000
     UNKNOWN = 0x100000000
+
+    @property
+    def collision(self):
+        """Returns `True` if the positioner is collided."""
+
+        return True if (PositionerStatus.ALPHA_COLLISION & self or
+                        PositionerStatus.BETA_COLLISION & self) else False
 
 
 class BootloaderStatus(Maskbit):